Promote Multipart to a top-level type with a nested builder.
This creates a model-like API in which individual parts can be read. It also offers the posibility of using this model to parse multipart in the future.
